Refactor subscription service configuration and update dependencies
- Renamed MySQL service from 'mysql' to 'mysql-db' in docker-compose.yml for clarity.
- Updated Spring DataSource URLs in multiple services to reflect the new MySQL service name.
- Added a new subscription service to docker-compose with appropriate configurations.
- Changed the exposed port in the subscription service Dockerfile from 8083 to 8087.
- Upgraded Spring Boot version in pom.xml from 3.5.0 to 3.5.3.
- Enabled JPA auditing in SubscriptionServiceApplication.java.
- Modified AuditableAbstractAggregateRoot and AuditableModel to remove 'updatable' constraint on createdAt field.
- Updated SubscriptionCommandServiceImpl to improve event publishing logic.
- Enhanced SubscriptionExpirationScheduler and SubscriptionMaintenanceScheduler for better logging and error handling.
- Changed price type in SubscriptionPlan, SubscriptionCreatedEvent, SubscriptionRenewedEvent, and SubscriptionType from Double to BigDecimal for precision.
- Removed deprecated application-docker.properties file and consolidated properties into application.properties.
- Updated OpenApiConfiguration and SecurityConfig for improved API documentation and security settings.

+++ subscription-service/src/main/java/com/hidrogreen/subscription_service/subscriptions/domain/model/entities/SubscriptionPlan.java
@@ -1,5 +1,7 @@
 package com.hidrogreen.subscription_service.subscriptions.domain.model.entities;
 
+import java.math.BigDecimal;
+
 import com.hidrogreen.subscription_service.shared.domain.model.entities.AuditableModel;
 import com.hidrogreen.subscription_service.subscriptions.domain.model.valueobjects.SubscriptionType;
 import jakarta.persistence.*;
@@ -36,7 +38,7 @@ public class SubscriptionPlan extends AuditableModel {
     @Getter
     @Setter
     @Column(name = "price", nullable = false, precision = 10, scale = 2)
-    private Double price;
+    private BigDecimal price;
 
     @Getter
     @Setter
@@ -69,7 +71,7 @@ public class SubscriptionPlan extends AuditableModel {
     private Boolean isActive;
 
     public SubscriptionPlan(SubscriptionType planType, String name, String description, 
-                           Double price, Integer durationDays, Integer maxCrops, 
+                           BigDecimal price, Integer durationDays, Integer maxCrops, 
                            Integer maxReports, Boolean hasPrioritySupport, 
                            Boolean hasAdvancedAnalytics, Boolean isActive) {
         this.planType = planType;
