+++ subscription-service/src/main/java/com/hidrogreen/subscription_service/subscriptions/domain/model/events/SubscriptionCreatedEvent.java
@@ -3,6 +3,7 @@ package com.hidrogreen.subscription_service.subscriptions.domain.model.events;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -12,12 +13,12 @@ public final class SubscriptionCreatedEvent extends ApplicationEvent {
     private final Long userId;
     private final String subscriptionType;
     private final String planName;
-    private final Double price;
+    private final BigDecimal price;
     private final String userEmail;
     private final LocalDateTime eventTime;
     
     public SubscriptionCreatedEvent(Object source, Long subscriptionId, Long userId, 
-                                  String subscriptionType, String planName, Double price, String userEmail) {
+                                  String subscriptionType, String planName, BigDecimal price, String userEmail) {
         super(source);
         this.subscriptionId = subscriptionId;
         this.userId = userId;
