+++ subscription-service/src/main/java/com/hidrogreen/subscription_service/subscriptions/domain/model/events/SubscriptionRenewedEvent.java
@@ -3,6 +3,7 @@ package com.hidrogreen.subscription_service.subscriptions.domain.model.events;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,13 +14,13 @@ public final class SubscriptionRenewedEvent extends ApplicationEvent {
     private final String oldSubscriptionType;
     private final String newSubscriptionType;
     private final String planName;
-    private final Double price;
+    private final BigDecimal price;
     private final String userEmail;
     private final LocalDateTime eventTime;
     
     public SubscriptionRenewedEvent(Object source, Long subscriptionId, Long userId, 
                                   String oldSubscriptionType, String newSubscriptionType,
-                                  String planName, Double price, String userEmail) {
+                                  String planName, BigDecimal price, String userEmail) {
         super(source);
         this.subscriptionId = subscriptionId;
         this.userId = userId;
