+++ subscription-service/src/main/java/com/hidrogreen/subscription_service/subscriptions/domain/model/valueobjects/SubscriptionType.java
@@ -1,17 +1,19 @@
 package com.hidrogreen.subscription_service.subscriptions.domain.model.valueobjects;
 
+import java.math.BigDecimal;
+
 public enum SubscriptionType {
-    FREE("Free Plan", 0.0, 30, "Basic features with limited access"),
-    BASIC("Basic Plan", 9.99, 30, "Enhanced features for individual farmers"),
-    PREMIUM("Premium Plan", 19.99, 30, "Advanced features with priority support"),
-    ENTERPRISE("Enterprise Plan", 49.99, 30, "Full access with custom integrations");
+    FREE("Free Plan", BigDecimal.valueOf(0.00), 30, "Basic features with limited access"),
+    BASIC("Basic Plan", BigDecimal.valueOf(9.99), 30, "Enhanced features for individual farmers"),
+    PREMIUM("Premium Plan", BigDecimal.valueOf(19.99), 30, "Advanced features with priority support"),
+    ENTERPRISE("Enterprise Plan", BigDecimal.valueOf(49.99), 30, "Full access with custom integrations");
 
     private final String displayName;
-    private final Double price;
+    private final BigDecimal price;
     private final Integer durationDays;
     private final String description;
 
-    SubscriptionType(String displayName, Double price, Integer durationDays, String description) {
+    SubscriptionType(String displayName, BigDecimal price, Integer durationDays, String description) {
         this.displayName = displayName;
         this.price = price;
         this.durationDays = durationDays;
@@ -22,7 +24,7 @@ public enum SubscriptionType {
         return displayName;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
